Handle Moving Cards
Made it so that the the number field for a card can be updated and the
card will move to that position in the list. If the enter key is pressed
after it's been updated, it will move the card and keep the focus on the
number field. If the number hasn't been changed and the enter key is
pressed, it will toggle the card out of focus. I think this is pretty
intuitive, right?