Add "Add Card" Functionality
Adding the Add Card functionality. Also changing the key for the CardList
cards to be the card's number, since it'll always be unique and I didn't
want to have to figure out what the next id would be.
I don't think that the way I'm setting the default type for the card
is very elegant, but it'll work for now...