Update CardList state
Added in the state updates in the CardList component for the deck name
and description. Looking at the state in the console, also observed
that the card.name values weren't updating in the state, which would be
bad if we saved and it was using the old value, as opposed to what was
showing up in the DOM. I don't think it's quiiiiiite what it should be
(updating the card directly and not re-making the whole cards array),
but it's good enough for now.