Check if Deck Exists
Before saving a deck, checking if the name is already taken. Eventually,
I think that we'll want to check name/author combinations or somesuch,
but for now, just checking if the deck name is already taken. It's not,
uh, super great, but it'll work for now. Getting to the point where I
figure I can change it later, so I might as well just get 'er done so
I can play the game. I want to play.