Use LocalStorage
Saving the default decks to the local storage and then loading from the
local storage after that. Also added in the save deck functionality
that will "overwrite" the deck that has the same name. 'Cause we don't
want decks with the same name, says I.....it's not super good, but I'll
maybe make it better someday...