Test Storage
Testing a textbox and storage to see if it works the way I expect it to—
that is, when the app is closed, when it opens back up again it retains
the text that was inputted into the textbox. Seems to be working great!
And so does eslint and prettifier! Typescript is playing nicely, too.
I'm very pleased with this setup.