Create HoC for Saving
Created a Higher Order Component to handle saving objects to and from the
store, since I want to be able to apply it to lots of different objects
easily. It was a SUPER DUPER PAIN trying to get TypeScript to work.
Committing now because it looks like it's working and I don't want to
lose my progress....