Save PanResponderExample
Saving and loading the PanResponderExamples now. It seems to be working
really great in a normal use-case scenario, but if the app crashes or
is forced to reload then it only has enough time to save one or two of
the component's data into storage before it restarts. So going to save
the state of the object into the store after every interaction so that
it retains state even if it crashes for the user. Looked up performance
and with this little data being written, I think we'll be more than fine...