Add Card Component
Added the base for the Card component. Looks a nice shape and can be
reversed upside-down. Had to get my hubby's help to figure out that the
z-axis was the correct one to rotate around to get the bottom corner number
to flip, because the x-axis was flipping and mirroring it. He's so wonderful
and handsome. I love him so much.