Put Cards on GameBoard
Hard coded the fantasy deck I want to play with and created Cards from it
and put them all on the GameBoard. Reordering the deck when a card is
selected so that it's drawn last and will appear on top of all of the
other cards. I'm not going to lie: I don't fully grock why it works.
I would have thought the deck.map in the GameBoard render would create
new Card components whenever the state changes, but I guess that React
is smart enough that when it sees that the keys and props haven't changed
it just re-orders the components instead of creating new ones?