Add UpsideDown Text
Added some right-side-up text when the cards are upsideDown so that it's
apparent that they're upsideDown but you can still read the card without
having to turn your whole phone upsideDown. Or, I guess laptop screen
if you were using your browser, hahahah....I'd like to make it so that
you could edit the grey italic text to whatever your interpretation of
what the upsideDown meaning is, but that's a feature for waaaaay later.