Mark Cards
Making it so that cards can be marked when they're doubleTapped on, so
that we'll be able to mark which ones shouldn't be shuffled back into
the deck when you get a reshuffle card or whatever. Had to make it so that
the cards weren't ACTUALLY upside down or the mark in the corner would
be in the wrong one. Oops. But we can always change it later if we need to...