Reset Cards (Except Marked)
Created a Start Game button that will move all of the cards back into
the deck and shuffle it, except for marked cards. This lets us keep cards
out for Character Facets, and later for our hands. Needed to update the
Card Component to update the styles of the card whenever, though, or
it wouldn't reset the card's position since those are more being tracked
by the styles and the underlying component, so....I think this is fine?
Seems to not have any performance issues on my phone....