Add Reshuffles
Adds reshuffle cards to the deck when we start a game. Also changes the
Start Game button to Reshuffle after we've started the game and added
the reshuffle cards to the deck. We're leaving it up to the user to
press the button to reshuffle when they come up, 'cause that's easiest
right now, hardy har har.