Save Game
Making it so that we save the game to the storage now every time a card
has finished moving. This makes it so that when you open the app back up
it has saved your spot. HooRay! Figure that the whole game is small enough
that we probably don't have to worry about performance issues or
anything...we won't worry about those until they actually become an issue.
Wonder how it will perform on older devices. Also added a touchable text
to the background that will clear the save and let you start a new game.
Seems to be working fine?