Add Loading Spinner
Adding a loading spinner for when it's pulling out the save state. The
amount of time it takes AsyncStorage.getItem to get back and render the
cards on the screen is super noticeable, so adding in the spinner and
only showing the GameBoard when we're reading makes it look a lot more
profreshional. Also changed the background to white, so you can see the
spinner better.