Don't Reorder Marked Cards
Making it so that if cards are marked then they won't get reshuffled into
the deck because that makes it annoying when you've set up your cards
a certain way and then all the sudden different ones than you expect are
on top or bottom. So taking them out of the deck before we reshuffle them.