Animate Cards
Turned the cards into Animated Views so that they can be moved around
and look real nice. It looks like it's properly saving and animating
and stuff...the trick was using this.moveAnimation.extractOffset() and
flattenOffset(). And also grabbing the values directly from the Animated.
Values from the getLayout()...that one feels really hacky...but it seems
to work on web and Android! Got the most help from this article:
https://dev.to/hyra/getting-started-with-the-panresponder-in-react-native-9mf