Add Deck and Update Deck Meta
Added deck meta data so that we can display the name of the deck we're
currently playing with. Also added another deck that we can play with
and made it so that we can select which one we want to play with
when we start a new game. Also added in a ton of comments for documentation.