Make Cards a Bit Shorter
Making the Card Components shorter so that we can fit all of the Facet
cards nicely in the top right corner without taking up as much real-estate
on the phone. Makes it easier to fit all of the cards on the screen
at once in a way where you can still read them without too much overlap.
All the text seems to fit pretty well, too, even when they're flipped.
Checked all the cards in both decks, and there's no weird overflow.
They'll probably need to be a bit bigger for when people can make their
own cards...